+++ Changelog.md
@@ -2,7 +2,9 @@
 
 ## v.12.4.0
 
-- Now option to surpress chat notifications on card shuffle
+- Now option to surpress chat notifications on card shuffle.
+- custom attacks are now possible on threats [471](https://github.com/gmmatt/torgeternity/issues/471)
+- A new macro that deals damage to targeted tokens [472](https://github.com/gmmatt/torgeternity/issues/472)
 
 ## v.12.3.1